refactor(ui): extract NodeDevice component from commonalities in PCI/USB components
