+++ ui/src/app/machines/views/MachineDetails/NodeDevices/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NodeDevices";
